修改字段类型，实现请求的自定义编码

+++ IM/Core/Utils/HETaskIdGenerator.h
@@ -11,5 +11,5 @@
 
 @interface HETaskIdGenerator : NSObject
 +(instancetype) shareInstance;
-- (NSUInteger)createId;
+- (uint32_t)createId;
 @end
